🐛 fix(lint): 修复 187 项 ruff 代码检查问题
- 自动修复 148 项（import 排序、未使用导入、不必要赋值等）
- 自动修复 33 项 unsafe-fixes（类型注解引号、未使用变量等）
- 手动修复 6 项：
  - os.path → Path（ocr_service.py, qrcode_service.py）
  - raise ... from None（ocr_service.py）
  - numpy → np 类型注解（ocr_service.py）
  - 添加 # noqa: E402 注释（延迟导入不可避免）
  - 存储 asyncio.ensure_future 引用（about_tab.py）
  - 导入 QPixmap 替代未定义的 PySide6 引用（screen_coordinate_mapper.py）

+++ tests/widgets/test_inline_canvas_mapper.py
@@ -1,4 +1,3 @@
-import pytest
 from PySide6.QtCore import QPoint, QPointF, QRect
 from PySide6.QtGui import QPixmap
 
@@ -39,7 +38,7 @@ class TestCanvasAcceptsMapper:
     def test_update_crop_region_with_mapper(self, qapp):
         canvas = InlineEditCanvas()
         mapper = _make_mapper(dpr=2.0)
-        screen_pixmap = QPixmap(int(200 * 2), int(100 * 2))
+        screen_pixmap = QPixmap((200 * 2), (100 * 2))
         screen_pixmap.setDevicePixelRatio(2.0)
         canvas.set_background(screen_pixmap, QPointF(0, 0), mapper)
         # Should not crash
